Stats calc, system messages, damage calc

+++ gameserver/src/main/java/com/shnok/javaserver/dto/external/serverpackets/ObjectMoveToPacket.java
@@ -5,14 +5,14 @@ import com.shnok.javaserver.enums.network.packettypes.external.ServerPacketType;
 import com.shnok.javaserver.model.Point3D;
 
 public class ObjectMoveToPacket extends SendablePacket {
-    public ObjectMoveToPacket(int id, Point3D pos, int speed, boolean walking) {
+    public ObjectMoveToPacket(int id, Point3D pos, float speed, boolean walking) {
         super(ServerPacketType.ObjectMoveTo.getValue());
 
         writeI(id);
         writeF(pos.getX());
         writeF(pos.getY());
         writeF(pos.getZ());
-        writeI(speed);
+        writeI((int) speed);
         writeB(walking ? (byte) 1 : (byte) 0);
         buildPacket();
     }
